Fixed a NumberFormatException bug introduced by using Long.valueOf instead of Double.valueOf.

+++ hail-yes-core/src/main/java/com/quane/hail/yes/SimpleLocation.java
@@ -43,7 +43,8 @@ public class SimpleLocation {
 	}
 
 	public void setLatitude(double latitude) {
-		this.latitude = Long.valueOf(preciseLocationFormatter.format(latitude));
+		this.latitude = Double.valueOf(preciseLocationFormatter
+				.format(latitude));
 	}
 
 	public double getLongitude() {
@@ -51,7 +52,7 @@ public class SimpleLocation {
 	}
 
 	public void setLongitude(double longitude) {
-		this.longitude = Long.valueOf(preciseLocationFormatter
+		this.longitude = Double.valueOf(preciseLocationFormatter
 				.format(longitude));
 	}
